Move openReport to VNet context
